fix: keep the QR encoder chunk-free on web-family builds (OK-59643)
